requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-extensions==4.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.8.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.3.0
sqlparse==0.5.4
whitenoise==6.11.0
